Rename command modules
