+++ src/main/java/duke/command/system/ByeCommand.java
@@ -1,4 +1,4 @@
-package duke.command.systemcommand;
+package duke.command.system;
 
 import duke.DukeException;
